Вынесла в модули работу с куками и таблицей
